Move validate_non_transaction_error_event into validators directory
